Unifying code for the unit-tests and the cli. Limiting the test code, so that it does not go above a memory limit.
