+++ test/BCFKeyValueData.cc
@@ -1,30 +1,11 @@
 #include <iostream>
 #include <map>
 #include "BCFKeyValueData.h"
+#include "BCFSerialize.h"
+#include "compare_iter.h"
 #include "catch.hpp"
 using namespace std;
 using namespace GLnexus;
-
-#define PSEUDO_RAND_SEED (1103)
-
-// generate a random number in the range [0 .. n-1]
-static int genRandNumber(int n){
-    static bool firstTime = true;
-
-    // initialization
-    if (firstTime) {
-        firstTime = false;
-        srand(PSEUDO_RAND_SEED);
-    }
-
-    int i = rand() % n;
-    return i;
-}
-
-static double genRandDouble(int n) {
-    return double(genRandNumber(n)) / double(n);
-}
-
 
 // Trivial in-memory KeyValue implementation used in unit tests.
 namespace KeyValueMem {
@@ -871,114 +852,6 @@ TEST_CASE("BCFKeyValueData::sampleset_range") {
     // TODO: need a test with only a subset of the samples
 }
 
-static void print_bcf_record(bcf1_t *x) {
-    cout << "rid=" << x->rid << endl;
-    cout << "pos=" << x->pos << endl;
-    cout << "rlen=" << x->rlen << endl;
-    cout << "qual=" << x->qual << endl;
-    cout << "n_info=" << x->n_info << endl;
-    cout << "n_allele=" << x->n_allele << endl;
-    cout << "n_sample=" << x->n_sample << endl;
-    cout << "shared.l=" << x->shared.l << endl;
-    cout << "indiv.l=" << x->indiv.l << endl;
-//    cout << std::string(x->indiv.s) << endl;
-}
-
-// return 1 if the records are the same, 0 otherwise
-static int bcf_shallow_compare(const bcf1_t *x, const bcf1_t *y) {
-    if (x->rid != y->rid) return 0;
-    if (x->pos != y->pos) return 0;
-    if (x->rlen != y->rlen) return 0;
-
-    // I think nan != nan, so we are skipping this comparison
-    //if (x->qual != y->qual) return 0;
-
-    if (x->n_info != y->n_info) return 0;
-    if (x->n_allele != y->n_allele) return 0;
-    if (x->n_sample != y->n_sample) return 0;
-    if (x->shared.l != y->shared.l) return 0;
-    if (x->indiv.l != y->indiv.l) return 0;
-
-    return 1;
-}
-
-using IterResults = map<string, shared_ptr<vector<shared_ptr<bcf1_t>>>>;
-
-static void read_entire_iter(RangeBCFIterator *iter, shared_ptr<IterResults> results) {
-    Status s;
-
-    string dataset;
-    shared_ptr<const bcf_hdr_t> hdr;
-    do {
-        vector<shared_ptr<bcf1_t>> v;
-        v.clear();
-        s = iter->next(dataset, hdr, v);
-        for (auto rec : v) {
-            if (results->find(dataset) == results->end()) {
-                auto w = make_shared<vector<shared_ptr<bcf1_t>>>();
-                (*results)[dataset] = w;
-            }
-            results->at(dataset)->push_back(rec);
-        }
-    } while (s.ok());
-}
-
-static void compare_results(shared_ptr<IterResults> resultsBase,
-                            shared_ptr<IterResults> resultsSoph) {
-    for(IterResults::iterator iter = resultsBase->begin();
-        iter != resultsBase->end();
-        ++iter) {
-        string dataset = iter->first;
-        shared_ptr<vector<shared_ptr<bcf1_t>>> v = iter->second;
-        shared_ptr<vector<shared_ptr<bcf1_t>>> w = resultsSoph->at(dataset);
-
-        REQUIRE(v->size() == w->size());
-        int len = v->size();
-        //cout << dataset << " len=" << len << endl;
-        for (int i=0; i < len; i++) {
-            if (bcf_shallow_compare((*v)[i].get(), (*w)[i].get()) == 0) {
-                cout << "Error comparing two records"  << endl;
-                print_bcf_record((*v)[i].get());
-                print_bcf_record((*w)[i].get());
-                REQUIRE(false);
-            }
-        }
-    }
-}
-
-// Read all the records in the range, and return as a vector of BCF records.
-// This is not a scalable function, because the return vector could be very
-// large. For debugging/testing use only.
-static void compare_query(T &data, MetadataCache &cache,
-                          const std::string& sampleset, const range& rng) {
-    //cout << "compare_query " << rng.str() << endl;
-
-    Status s;
-    vector<unique_ptr<RangeBCFIterator>> iterators;
-    shared_ptr<const set<string>> samples, datasets;
-
-    // simple iterator
-    REQUIRE(data.sampleset_range_base(cache, sampleset, rng,
-                                      samples, datasets, iterators).ok());
-
-    auto resultsBase = make_shared<IterResults>();
-    for (int i=0; i < iterators.size(); i++) {
-        read_entire_iter(iterators[i].get(), resultsBase);
-    }
-    iterators.clear();
-
-    // sophisticated iterator
-    auto resultsSoph = make_shared<IterResults>();
-    REQUIRE(data.sampleset_range(cache, sampleset, rng,
-                                 samples, datasets, iterators).ok());
-    for (int i=0; i < iterators.size(); i++) {
-        read_entire_iter(iterators[i].get(), resultsSoph);
-    }
-
-    // compare all the elements between the two results
-    compare_results(resultsBase, resultsSoph);
-}
-
 TEST_CASE("BCFKeyValueData compare iterator implementations") {
     // This tests the optimized bucket-based range slicing in BCFKeyValueData
     int nRegions = 13;
@@ -1015,7 +888,12 @@ TEST_CASE("BCFKeyValueData compare iterator implementations") {
         if (end < beg)
             std::swap(beg, end);
         range rng(0, beg, end);
-        compare_query(*data, *cache, sampleset, rng);
+        int rc = compare_query(*data, *cache, sampleset, rng);
+        switch (rc) {
+        case 1: break; // comparison succeeded
+        case 0: REQUIRE(false); // ERROR
+        case -1: break;  // Query used too much memory, continue
+        }
     }
 
     cout << "Compared " << nIter << " range queries between the two iterators" << endl;
